More work on getting Common to compile.
Testing on Linux, vtkBox is one major remaining issue in DataModel - it
is an implicit function, and one static method is used in vtkCell3D.
This can probably just be refactored as it is internal.

Still outstanding it the vtkInformation issue, currently commented out
in Core to to dependencies on vtkDataObject in one key type.

Change-Id: Id0a5f59bb29705d0d5c4d812397d1413e3b5590e