Simplify vtkAutoInit macros
The previous implementation dependend on preprocessors capable of
handling the empty string as a macro argument during expansion.  Not all
compilers are consistent.  Replace use of arbitrary nesting of
lisp-style lists with a simple tuple approach.

Use a tuple of the form "#(...)" where "#" is the number of arguments in
"...".  For example:

  '-DvtkRenderingCore_AUTOINIT=1(vtkRenderingOpenGL)'
  '-DvtkIOSQL_AUTOINIT=2(vtkIOMySQL,vtkIOPostgreSQL)'

This tuple format reads well and is simple to handle with preprocessor
symbol concatenation.